Добавление типов, исправления наименований
