+++ src/components/app/App.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect }  from 'react';
 import styles from './App.module.css';
 import AppHeader from '../app-header/app-header.component';
-import BurgerConstructor from '../burger-constructor/burger-constructor.component';
 import BurgerIngredients from '../burger-ingredients/burger-ingredients.component';
+import BurgerConstructor from '../burger-constructor/burger-constructor.component';
 import { data, dataUrl } from '../../data';
 
 function App() 
@@ -28,9 +28,9 @@ function App()
         <AppHeader />
         <div className={styles.main_control} >
           <div className={`mr-10`}>
-            <BurgerConstructor allData={loadedData} />
+            <BurgerIngredients allData={loadedData} />
           </div>
-          <BurgerIngredients allData={loadedData} />
+          <BurgerConstructor allData={loadedData} />
         </div>
     </div>
   );
