+++ src/components/burger-constructor/burger-constructor.styles.module.css
@@ -1,6 +1,17 @@
+.checkout_st {
+    display: flex; 
+    flex-direction: row; 
+    justify-content: flex-end;
+    gap: 10px
+}
 
+.panel {
+    display: flex; 
+    flex-direction: column;
+}
 
-
-
-
-
+.menu {
+    display: flex; 
+    flex-direction: column; 
+    gap: 10px; 
+}
